refactor(Types): ♻️ Moved timeline & date type definitions.     -	Migrated timeline-related interfaces from `props.d.ts` to a     dedicated `timeline.d.ts` file.
    -	Extracted date-specific types into `date.d.ts` from both
    `timeline.d.ts` and `props.d.ts`.

    -	Updated all import references to ensure consistency with the new
    `timeline.d.ts` structure. Isolated temporal logic into domain-
    specific modules to improve codebase maintainability.

    -	Reduced the complexity of the global `props.d.ts` by delegating
    responsibilities to dedicated files.
