+++ frontend/components/timeline/Card.vue
@@ -29,7 +29,7 @@
                 </section>
                 <section v-if="!!data.techStack && data.techStack.length > 0"
                     :class="['tech-container']">
-                    <h4>Teknologi : </h4>
+                    <h4>Teknologi(er) : </h4>
                     <p :class="['flex-wrap-row-justify-space-evenly']">
                         <span v-for="(tech, i) in data.techStack" :key="i">
                             <MediaFigure v-if="tech" :data="tech"
