+++ frontend/types/timeline.d.ts
@@ -1,6 +1,6 @@
 //  Types for the project
-
-import type { Anchor} from "./props";
+import type { DateObject } from "./date";
+import type { Anchor} from "./navigation";
 
 export interface TimelineItem
 {
@@ -24,12 +24,8 @@ export interface TimelineItem
     }
 }
 
-export interface DateObject
-{
-    created: string;
-    end?: string | null;
-    updated?: string | null;
-}
-
 export interface TechStack { category: string; label: string; }
 export interface ReferencePoint { name: string; anchor: Anchor; }
+export interface FilterProps { cls?: Array<any>; data: Record<string, any>; }
+export interface TimelineCardProps { cls?: Array<any>; data: TimelineItem; isVisible?: boolean; }
+export interface TimelineProps { title: string; range?: number; cls?: Array<any>; data: TimelineItem[]; }
